# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hostreamly Media API",
    "version": "2.0.0",
    "description": "API completa para la plataforma Hostreamly Media con funcionalidades empresariales avanzadas. Esta API proporciona endpoints para gestión de videos, streaming en vivo, analíticas avanzadas, webhooks personalizados, formularios de contacto, enlaces de descarga controlados y mucho más.",
    "termsOfService": "https://hostreamly.com/terms",
    "contact": {
      "name": "Hostreamly Support",
      "email": "support@hostreamly.com",
      "url": "https://hostreamly.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Servidor de desarrollo"
    },
    {
      "url": "https://api.hostreamly.com/api",
      "description": "Servidor de producción"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/enterprise/sla/alerts": {
      "get": {
        "summary": "Obtener alertas de SLA",
        "tags": ["Enterprise - SLA"],
        "responses": {
          "200": {
            "description": "Lista de alertas de SLA"
          }
        }
      },
      "post": {
        "summary": "Crear nueva alerta de SLA",
        "tags": ["Enterprise - SLA"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "metric": {
                    "type": "string"
                  },
                  "threshold": {
                    "type": "number"
                  },
                  "condition": {
                    "type": "string",
                    "enum": ["above", "below"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Alerta creada exitosamente"
          }
        }
      }
    },
...
    "/enterprise/account/team/invite": {
      "post": {
        "summary": "Invitar miembro al equipo",
        "tags": ["Enterprise - Gestión de Cuenta"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "role": {
                    "type": "string",
                    "enum": ["admin", "manager", "user", "viewer"]
                  },
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invitación enviada exitosamente"
          }
        }
      }
    },
    "/enterprise/reports/generate": {
      "post": {
        "summary": "Generar reporte Enterprise",
        "tags": ["Enterprise - Reportes"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": ["sla_compliance", "account_usage", "team_activity", "security_audit", "performance_analysis"]
                  },
                  "format": {
                    "type": "string",
                    "enum": ["excel", "pdf", "csv", "json"]
                  },
                  "dateRange": {
                    "type": "object",
                    "properties": {
                      "start": {
                        "type": "string",
                        "format": "date"
                      },
                      "end": {
                        "type": "string",
                        "format": "date"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reporte generado exitosamente"
          }
        }
      }
    }
  },
...
}